PDF_ READER_ USING KB/test.py
from neo4j import GraphDatabase
import os
from dotenv import load_dotenv

load_dotenv()

uri = os.getenv("NEO4J_URI", "bolt://localhost:7687")
user = os.getenv("NEO4J_USER", "neo4j")
password = os.getenv("NEO4J_PASSWORD", "")

try:
    driver = GraphDatabase.driver(uri, auth=(user, password))
    with driver.session() as session:
        result = session.run("RETURN 1 AS ok").single()
        print("✅ Neo4j connection successful!", result)
except Exception as e:
    print("❌ Failed to connect to Neo4j")
    print(e)
